Level Finish UI, fixed line on timer script, progress on level manager script
LevelFinishUI displays the  player's final score when Display() is called. It contains a LoadScene() void with a scene name string parameter. This void was ideally made to allow the player to load back into the main menu.

Line 25 was fixed on the 'Timer' script, timer should resume when 'pauseTime' is false now.

LevelManager has progress done. It includes a score system that can be referenced by other scripts through GetScore(). AddScore() can be used by other scripts to increase score value, and ResetScore() is to reset it back to zero when appropriate.